+++ cropper/src/main/java/com/canhub/cropper/CropImageOptions.kt
@@ -43,11 +43,13 @@ open class CropImageOptions : Parcelable {
     /** The shape of the cropping window.  */
     @JvmField
     var cropShape: CropShape
+
     /**
      * The shape of cropper corners
      */
     @JvmField
     var cornerShape: CropImageView.CropCornerShape
+
     /**
      * The radius of the circular crop corner
      */
@@ -167,6 +169,7 @@ open class CropImageOptions : Parcelable {
     @JvmField
     @ColorInt
     var borderCornerColor: Int
+
     /**
      * The fill color of circle corner
      */
@@ -234,6 +237,11 @@ open class CropImageOptions : Parcelable {
     @JvmField
     @ColorInt
     var activityMenuIconColor: Int
+
+    /** the color to use for action bar items texts  */
+    @JvmField
+    @ColorInt
+    var activityMenuTextColor: Int? = null
 
     /** the Android Uri to save the cropped image to  */
     @JvmField
@@ -349,7 +357,19 @@ open class CropImageOptions : Parcelable {
 
     /** Toolbar color **/
     @JvmField
-    var toolbarColor: Int = -1
+    var toolbarColor: Int? = null
+
+    /** Toolbar color **/
+    @JvmField
+    var toolbarTitleColor: Int? = null
+
+    /** Toolbar color **/
+    @JvmField
+    var toolbarBackButtonColor: Int? = null
+
+    /** Toolbar tint color **/
+    @JvmField
+    var toolbarTintColor: Int? = null
 
     /** Init options with defaults.  */
     constructor() {
@@ -393,6 +413,7 @@ open class CropImageOptions : Parcelable {
         maxCropResultHeight = 99999
         activityTitle = ""
         activityMenuIconColor = 0
+        activityMenuTextColor = null
         customOutputUri = null
         outputCompressFormat = CompressFormat.JPEG
         outputCompressQuality = 90
@@ -418,7 +439,10 @@ open class CropImageOptions : Parcelable {
         cropperLabelTextColor = Color.WHITE
         showCropLabel = false
         activityBackgroundColor = Color.WHITE
-        toolbarColor = -1
+        toolbarColor = null
+        toolbarTitleColor = null
+        toolbarBackButtonColor = null
+        toolbarTintColor = null
     }
 
     /** Create object from parcel.  */
@@ -461,6 +485,7 @@ open class CropImageOptions : Parcelable {
         maxCropResultHeight = parcel.readInt()
         activityTitle = TextUtils.CHAR_SEQUENCE_CREATOR.createFromParcel(parcel)
         activityMenuIconColor = parcel.readInt()
+        activityMenuTextColor = parcel.readValue(Int::class.java.classLoader) as Int?
         customOutputUri = parcel.readParcelable(Uri::class.java.classLoader)
         outputCompressFormat = CompressFormat.valueOf(parcel.readString()!!)
         outputCompressQuality = parcel.readInt()
@@ -487,7 +512,10 @@ open class CropImageOptions : Parcelable {
         cropperLabelText = parcel.readString()!!
         showCropLabel = parcel.readByte().toInt() != 0
         activityBackgroundColor = parcel.readInt()
-        toolbarColor = parcel.readInt()
+        toolbarColor = parcel.readValue(Int::class.java.classLoader) as Int?
+        toolbarTitleColor = parcel.readValue(Int::class.java.classLoader) as Int?
+        toolbarBackButtonColor = parcel.readValue(Int::class.java.classLoader) as Int?
+        toolbarTintColor = parcel.readValue(Int::class.java.classLoader) as Int?
     }
 
     override fun writeToParcel(dest: Parcel, flags: Int) {
@@ -529,6 +557,7 @@ open class CropImageOptions : Parcelable {
         dest.writeInt(maxCropResultHeight)
         TextUtils.writeToParcel(activityTitle, dest, flags)
         dest.writeInt(activityMenuIconColor)
+        dest.writeValue(activityMenuTextColor)
         dest.writeParcelable(customOutputUri, flags)
         dest.writeString(outputCompressFormat.name)
         dest.writeInt(outputCompressQuality)
@@ -555,7 +584,10 @@ open class CropImageOptions : Parcelable {
         dest.writeString(cropperLabelText)
         dest.writeByte((if (showCropLabel) 1 else 0).toByte())
         dest.writeInt(activityBackgroundColor)
-        dest.writeInt(toolbarColor)
+        dest.writeValue(toolbarColor)
+        dest.writeValue(toolbarTitleColor)
+        dest.writeValue(toolbarBackButtonColor)
+        dest.writeValue(toolbarTintColor)
     }
 
     override fun describeContents(): Int {